requirements.txt
pandas
openpyxl
psycopg2-binary
pyyaml
numpy
xlrd
boto3
fsspec
s3fs
pytest
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-xdist
Jinja2
pyarrow
pandas-schema
pandera
pytest-env
